We should use a size_t for rangeCount. We do some arithmetic internally with sizeof's. We'll hit allocation troubles before running out of values.
git-svn-id: svn+ssh://svn.phoenixbox.net/svn/v6502/trunk@754 96dfbd92-d197-e211-9ca9-00110a534b34

+++ v6502/mem.c
@@ -29,7 +29,7 @@
 #pragma mark Memory Lifecycle
 
 v6502_mappedRange *v6502_mappedRangeForOffset(v6502_memory *memory, uint16_t offset) {
-	for (int i = 0; i < memory->rangeCount; i++) {
+	for (size_t i = 0; i < memory->rangeCount; i++) {
 		v6502_mappedRange *currentRange = &memory->mappedRanges[i];
 		if (offset >= currentRange->start && offset < (currentRange->start + currentRange->size)) {
 			return currentRange;
